resolve commits
